Enable viewing taskid in Plain view to help edit
So far, TASKID wasn't shown in Plain viewing mode, so that one would have to call --format=org in order to edit any events (or mark them completed by task_id, etc.). This simple change adds the taskid at the beginning of each entry in plain text. I find it to be unobtrusive and very helpful.